Move related types into utils/compountMeasures
