+++ src/utils/compoundMeasures.ts
@@ -1,11 +1,24 @@
-import {
-  COMPOUND_MEASURE_DELIMITER,
-  CompoundMeasure,
-  CompoundMeasureType,
-  PeriodOverPeriodMeasure,
-  RollingMeasure,
-} from '../services/data/types';
+export enum CompoundMeasureType {
+  ROLLING,
+  PERIOD_OVER_PERIOD,
+}
 
+export interface CompoundMeasure {
+  originalMeasureName: string;
+  type: CompoundMeasureType;
+}
+
+export interface RollingMeasure extends CompoundMeasure {
+  type: CompoundMeasureType.ROLLING;
+  windowSize: number;
+}
+
+export interface PeriodOverPeriodMeasure extends CompoundMeasure {
+  type: CompoundMeasureType.PERIOD_OVER_PERIOD;
+  periodOffset: number;
+}
+
+export const COMPOUND_MEASURE_DELIMITER = '$$';
 
 export function isCompoundMeasure(measureName: string) {
   return measureName.includes(COMPOUND_MEASURE_DELIMITER);
